chore: Update dependencies in requirements.txt
The dependencies in requirements.txt have been updated to their latest versions. This includes upgrading the `anthropic` package to version 0.32.0 and the `openai` package to version 1.40.2. Additionally, the `requests` package has been updated to version 2.32.3. The `transformers` package has been upgraded to version 4.44.0, the `datasets` package to version 2.20.0, and the `tiktoken` package to version 0.7.0. The `pymongo` package has been updated to version 4.8.0, and the `nltk` package to version 3.8.1. Lastly, the `redis` package has been updated to version 5.0.8. The `google-api-python-client` package has been upgraded to version 2.140.0, the `google-auth` package to version 2.33.0, the `google-auth-httplib2` package to version 0.2.0, and the `google-auth-oauthlib` package to version 1.2.1. The `serpapi` package has been updated to version 0.1.5.